Cleanup code, fix img url, fix empty space, etc
Changelog:
- Bebersih code (code cleanup)
- Daftar Artikel Terunduh dan Kajian Tersimpan sekarang sudah bisa update secara otomatis (TODO 20/08/2020)
- Terdapat konfirmasi saat mencoba untuk menghapus kajian tersimpan dan menghapus artikel yang telah diunduh
- Konversi Java ke Kotlin untuk class ArtikelActivity dan KajianActivity
- Sekarang Anda dapat me-refresh daftar artikel dan kajian dengan cara tarik list ke bawah layar.
- Memperbaiki URL pada ReadArticleActivity, karena selalu dianggap ada gambar tanpa alamat URL
- Mempercantik tampilan Lihat Kajian dan Lihat Artikel
- NIM Mar'Atun Solehah diperbaiki.

+++ ShoumHome-android/app/src/main/java/id/shoumhome/android/activity/ReadArticleActivity.kt
@@ -6,6 +6,7 @@ import android.view.Menu
 import android.view.MenuItem
 import android.view.View
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.lifecycle.Observer
@@ -70,6 +71,8 @@ class ReadArticleActivity : AppCompatActivity() {
                 tvLikeCount.text = likes
                 tvPostDate.text = it["post_date"].toString()
                 tvPostContent.text = it["content"].toString()
+                val ustadzName = resources.getString(R.string.by) + " " + it["ustadz_name"].toString()
+                tv_nama_Ustad.text = ustadzName
                 progressMessage.visibility = View.GONE
                 errorMessage.visibility = View.GONE
                 article = Article()
@@ -79,7 +82,7 @@ class ReadArticleActivity : AppCompatActivity() {
                 article.post_date = it["post_date"].toString()
                 article.ustadzName = it["ustadz_name"].toString()
                 article.hasImg = it["hasImg"].toString()
-                article.imgUrl = ""
+                article.imgUrl = resources.getString(R.string.server) + "assets/articles/${id}.${it["extension"]}"
                 article.likes = it["like"].toString().toInt()
             } else {
                 progressMessage.visibility = View.GONE
@@ -112,10 +115,19 @@ class ReadArticleActivity : AppCompatActivity() {
         }
         fabDownload.setOnClickListener {
             if(downloaded){
-                dbArticleHelper.deleteById(id)
-                downloaded= false
-                fabDownload.setImageResource(R.drawable.ic_baseline_assignment_returned_24)
-                Toast.makeText(this, "Artikel telah dihapus!",Toast.LENGTH_LONG).show()
+                val alert = AlertDialog.Builder(this)
+                alert.setTitle(resources.getString(R.string.sure))
+                alert.setMessage(resources.getString(R.string.remove_article_confirm))
+                alert.setPositiveButton(resources.getString(R.string.yes)) { _, _ ->
+                    dbArticleHelper.deleteById(id)
+                    downloaded = false
+                    fabDownload.setImageResource(R.drawable.ic_baseline_assignment_returned_24)
+                    Toast.makeText(this, "Artikel telah dihapus!", Toast.LENGTH_LONG).show()
+                }
+                alert.setNegativeButton(resources.getString(R.string.no)) { _, _ ->
+                    /* no-op */
+                }
+                alert.create().show()
             }else{
                 val values= ContentValues()
                 values.put(DatabaseContract.ArticleColums.ID, article.id)
@@ -129,6 +141,45 @@ class ReadArticleActivity : AppCompatActivity() {
                 downloaded= true
                 fabDownload.setImageResource(R.drawable.ic_baseline_assignment_turned_in_24)
                 Toast.makeText(this, "Artikel Berhasil Diunduh", Toast.LENGTH_LONG).show()
+            }
+        }
+        // pullToRefresh
+        pullToRefresh.setOnRefreshListener {
+            svArticle.visibility = View.GONE
+            pullToRefresh.isRefreshing = false
+            errorMessage.visibility = View.GONE
+            progressMessage.visibility = View.VISIBLE
+            GlobalScope.launch(Dispatchers.Main) {
+                val deferredArticles = async(Dispatchers.IO) {
+                    id.let { readArticleViewModel.setArticle(applicationContext, it) }
+                }
+                val it = deferredArticles.await()
+                if (it["status"] == true) {
+                    title = it["title"].toString()
+                    content = it["content"].toString()
+                    tvArticleTitle.text = it["title"].toString()
+                    if (it["hasImg"] == true) {
+                        Glide.with(applicationContext)
+                                .load(resources.getString(R.string.server) + "assets/articles/${id}.${it["extension"]}")
+                                .into(imgArticle)
+                    } else {
+                        imgArticle.visibility = View.GONE
+                    }
+                    val likes = it["like"].toString() + " ${resources.getString(R.string.likes)}"
+                    tvLikeCount.text = likes
+                    tvPostDate.text = it["post_date"].toString()
+                    tvPostContent.text = it["content"].toString()
+                    progressMessage.visibility = View.GONE
+                    svArticle.visibility = View.VISIBLE
+                } else {
+                    progressMessage.visibility = View.GONE
+                    errorMessage.visibility = View.VISIBLE
+                    val errorMessage = """
+                            Error!
+                            [${it["code"]}]: ${it["message"]}
+                        """.trimIndent()
+                    tvErrorMessage.text = errorMessage
+                }
             }
         }
     }
