+++ ShoumHome-android/app/src/main/java/id/shoumhome/android/activity/ReadArticleSQLActivity.kt
@@ -4,9 +4,12 @@ import android.content.ContentValues
 import android.os.Bundle
 import android.os.PersistableBundle
 import android.view.Menu
+import android.view.MenuItem
 import android.view.View
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
 import com.bumptech.glide.Glide
 import com.google.android.gms.common.internal.ConnectionErrorMessages
 import id.shoumhome.android.R
@@ -38,17 +41,29 @@ class ReadArticleSQLActivity: AppCompatActivity() {
         //menutup layar progres dan error
         progressMessage.visibility = View.GONE
         errorMessage.visibility = View.GONE
+        imgLike.visibility = View.GONE
+        tvLikeCount.visibility = View.GONE
 
         val bundle = intent.extras!!
-        article = bundle?.getParcelable<Article>(EXTRA_PARCEL_ARTICLES)!!
+        article = bundle.getParcelable(EXTRA_PARCEL_ARTICLES)!!
         tvArticleTitle.text = article.title
         tvPostDate.text = article.post_date
-        tv_nama_Ustad.text = article.ustadzName
+        val ustadzName = resources.getString(R.string.by) + " " + article.ustadzName
+        tv_nama_Ustad.text = ustadzName
         tvPostContent.text = article.content
-        Glide.with(this)
-                .load(article.imgUrl)
-                .into(imgArticle)
 
+        if (article.hasImg == "true") {
+            Glide.with(this)
+                    .load(article.imgUrl) //masih bug
+                    .into(imgArticle)
+        } else {
+            imgArticle.visibility = View.GONE
+        }
+
+        val toolbar = findViewById<Toolbar>(R.id.toolbar)
+        setSupportActionBar(toolbar)
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+        supportActionBar?.title = resources.getString(R.string.read_article)
 
         dbArticleHelper = DbArticleHelper.getInstance(this)
         dbArticleHelper.open()
@@ -70,10 +85,20 @@ class ReadArticleSQLActivity: AppCompatActivity() {
         }
         fabDownload.setOnClickListener {
             if (downloaded) {
-                article.id?.let { it1 -> dbArticleHelper.deleteById(it1) }
-                downloaded = false
-                fabDownload.setImageResource(R.drawable.ic_baseline_assignment_returned_24)
-                Toast.makeText(this, "Artikel telah dihapus!", Toast.LENGTH_LONG).show()
+                val alert = AlertDialog.Builder(this)
+                alert.setTitle(resources.getString(R.string.sure))
+                alert.setMessage(resources.getString(R.string.remove_article_confirm))
+                alert.setPositiveButton(resources.getString(R.string.yes)) { _, _ ->
+                    article.id?.let { it1 -> dbArticleHelper.deleteById(it1) }
+                    downloaded = false
+                    fabDownload.setImageResource(R.drawable.ic_baseline_assignment_returned_24)
+                    Toast.makeText(this, "Artikel telah dihapus!", Toast.LENGTH_LONG).show()
+                    finish()
+                }
+                alert.setNegativeButton(resources.getString(R.string.no)) { _, _ ->
+                    /* no-op */
+                }
+                alert.create().show()
             } else {
                 val values = ContentValues()
                 values.put(DatabaseContract.ArticleColums.ID, article.id)
@@ -88,8 +113,13 @@ class ReadArticleSQLActivity: AppCompatActivity() {
                 fabDownload.setImageResource(R.drawable.ic_baseline_assignment_turned_in_24)
                 Toast.makeText(this, "Artikel Berhasil Diunduh", Toast.LENGTH_LONG).show()
             }
-
+            pullToRefresh.setOnRefreshListener { pullToRefresh.isRefreshing = false }
         }
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        if (item.itemId == android.R.id.home) super.onBackPressed()
+        return true
     }
 }
